HW13 - SQLStorage refactor

+++ src/ru/javawebinar/basejava/storage/SqlStorage.java
@@ -1,19 +1,19 @@
 package ru.javawebinar.basejava.storage;
 
+import ru.javawebinar.basejava.exception.ExistStorageException;
 import ru.javawebinar.basejava.exception.NotExistStorageException;
 import ru.javawebinar.basejava.model.Resume;
-import ru.javawebinar.basejava.util.SqlHelper;
+import ru.javawebinar.basejava.sql.SqlHelper;
 
 import java.sql.*;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class SqlStorage implements Storage {
     public final SqlHelper helper;
 
-    public SqlStorage(SqlHelper helper) {
-        this.helper = helper;
+    public SqlStorage(String dbUrl, String dbUser, String dbPassword) {
+        this.helper = new SqlHelper(dbUrl, dbUser, dbPassword);
     }
 
     @Override
@@ -38,6 +38,9 @@ public class SqlStorage implements Storage {
 
     @Override
     public void update(Resume r) {
+        if (!isInDB(r.getUuid())) {
+            throw new NotExistStorageException(r.getUuid());
+        }
         helper.execute("UPDATE resume SET full_name = ? WHERE uuid = ?", (ps) -> {
             ps.setString(1, r.getFullName());
             ps.setString(2, r.getUuid());
@@ -48,6 +51,9 @@ public class SqlStorage implements Storage {
 
     @Override
     public void save(Resume r) {
+        if (isInDB(r.getUuid())) {
+            throw new ExistStorageException(r.getUuid());
+        }
         helper.execute("INSERT INTO resume (uuid, full_name) VALUES (?,?)", (ps) -> {
             ps.setString(1, r.getUuid());
             ps.setString(2, r.getFullName());
@@ -58,38 +64,46 @@ public class SqlStorage implements Storage {
 
     @Override
     public void delete(String uuid) {
+        if (!isInDB(uuid)) {
+            throw new NotExistStorageException(uuid);
+        }
         helper.execute("DELETE FROM resume r WHERE r.uuid =?", (ps) -> {
             ps.setString(1, uuid);
-            if (!ps.execute()) {
-                throw new NotExistStorageException(uuid);
-            }
+            ps.execute();
             return null;
         });
     }
 
     @Override
     public List<Resume> getAllSorted() {
-        final List<Resume> list = new ArrayList<>();
-        helper.execute("SELECT uuid FROM resume", (ps) -> {
+        return helper.execute("SELECT uuid FROM resume ORDER BY full_name", (ps) -> {
+            List<Resume> list = new ArrayList<>();
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
                 list.add(get(rs.getString("uuid")));
             }
-            Collections.sort(list);
-            return null;
+            return list;
         });
-        return list;
     }
 
     @Override
     public int size() {
-        final int[] i = new int[1];
-        helper.execute("SELECT count(*) FROM resume", (ps) -> {
+        return helper.execute("SELECT count(*) FROM resume", (ps) -> {
             ResultSet rs = ps.executeQuery();
             rs.next();
-            i[0] = Integer.parseInt(rs.getString(1));
-            return null;
+            return Integer.parseInt(rs.getString(1));
         });
-        return i[0];
+    }
+
+    private boolean isInDB(String uuid) {
+        return helper.execute("SELECT uuid FROM resume", (ps) -> {
+            List<String> list = new ArrayList<>();
+            ResultSet rs = ps.executeQuery();
+            while (rs.next()) {
+                list.add(rs.getString("uuid").trim());
+            }
+            list.forEach(System.out::println);
+            return list;
+        }).contains(uuid);
     }
 }
